move ToCommitStatus into its own file
